+++ packages/user-service/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.1.1
+
+- refactor: getUserInfo from command in O2
+- fix: set timeout for getUserInfoFromDefClient
+- fix: only run getUserInfoFromDefClient in AliInternal
+
 ## 0.1.0
 
 - feat: init